Remove OpenNotificationAction.
NOTE: not fully working software.

Resolves #332.
