+++ Sources/UI/Notifications/Services/NotificationHandler/NotificationHandler.swift
@@ -10,5 +10,5 @@ import UserNotifications
 
 public protocol NotificationHandler {
     @discardableResult func handle(_ response: UNNotificationResponse, completionHandler: (() -> Void)?) -> Bool
-    func action(for response: UNNotificationResponse) -> Action?
+    func notification(for response: UNNotificationResponse) -> Notification?
 }
