+++ Sources/UI/Notifications/Services/NotificationHandler/NotificationHandlerService.swift
@@ -7,6 +7,8 @@
 //
 
 import UserNotifications
+import UIKit
+import os
 
 class NotificationHandlerService: NotificationHandler {
     let dispatcher: Dispatcher
@@ -14,27 +16,58 @@ class NotificationHandlerService: NotificationHandler {
     
     typealias ActionProvider = (Notification) -> Action?
     let actionProvider: ActionProvider
+    let notificationStore: NotificationStore
+    let eventQueue: EventQueue
     
-    init(dispatcher: Dispatcher, influenceTracker: InfluenceTracker, actionProvider: @escaping ActionProvider) {
+    init(dispatcher: Dispatcher, influenceTracker: InfluenceTracker, actionProvider: @escaping ActionProvider, notificationStore: NotificationStore, eventQueue: EventQueue) {
         self.dispatcher = dispatcher
         self.actionProvider = actionProvider
         self.influenceTracker = influenceTracker
+        self.notificationStore = notificationStore
+        self.eventQueue = eventQueue
     }
+    
+    
     
     func handle(_ response: UNNotificationResponse, completionHandler: (() -> Void)?) -> Bool {
         // The app was opened directly from a push notification. Clear the last received
         // notification from the influence tracker so we don't erroneously track an influenced open.
         influenceTracker.clearLastReceivedNotification()
         
-        guard let action = action(for: response) else {
+        guard let notification = notification(for: response) else {
             return false
         }
         
-        dispatcher.dispatch(action, completionHandler: completionHandler)
+        notificationStore.addNotification(notification)
+        
+        if !notification.isRead {
+            notificationStore.markNotificationRead(notification.id)
+        }
+        
+        switch notification.tapBehavior {
+        case .openApp:
+            break
+        case .openURL(let url):
+            DispatchQueue.main.sync {
+                UIApplication.shared.open(url, options: [:], completionHandler: nil)
+            }
+        
+        case .presentWebsite(let url):
+            // TODO
+            os_log("TODO .presentWebsite currently NOT IMPLEMENTED.", log: .general, type: .error)
+//            if let action = presentWebsiteActionProvider(url) {
+//                produceAction(action)
+//            }
+        }
+        
+        let eventInfo = notification.openedEvent(source: .pushNotification)
+        eventQueue.addEvent(eventInfo)
+        completionHandler?()
+        
         return true
     }
     
-    func action(for response: UNNotificationResponse) -> Action? {
+    func notification(for response: UNNotificationResponse) -> Notification? {
         guard let data = try? JSONSerialization.data(withJSONObject: response.notification.request.content.userInfo, options: []) else {
             return nil
         }
@@ -49,7 +82,7 @@ class NotificationHandlerService: NotificationHandler {
         
         do {
             let payload = try JSONDecoder.default.decode(Payload.self, from: data)
-            return actionProvider(payload.rover.notification)
+            return payload.rover.notification
         } catch {
             return nil
         }
