+++ Sources/UI/UIAssembler.swift
@@ -187,21 +187,6 @@ extension UIAssembler: Assembler {
             )
         }
         
-        // MARK: Action (openNotification)
-        
-        container.register(Action.self, name: "openNotification", scope: .transient) { (resolver, notification: Notification) in
-            let presentWebsiteActionProvider: OpenNotificationAction.ActionProvider = { [weak resolver] url in
-                return resolver?.resolve(Action.self, name: "presentWebsite", arguments: url)!
-            }
-            
-            return OpenNotificationAction(
-                eventQueue: resolver.resolve(EventQueue.self)!,
-                notification: notification,
-                notificationStore: resolver.resolve(NotificationStore.self)!,
-                presentWebsiteActionProvider: presentWebsiteActionProvider
-            )
-        }
-        
         // MARK: Action (presentNotificationCenter)
         
         container.register(Action.self, name: "presentNotificationCenter", scope: .transient) { resolver in
@@ -242,7 +227,9 @@ extension UIAssembler: Assembler {
             return NotificationHandlerService(
                 dispatcher: resolver.resolve(Dispatcher.self)!,
                 influenceTracker: resolver.resolve(InfluenceTracker.self)!,
-                actionProvider: actionProvider
+                actionProvider: actionProvider,
+                notificationStore: resolver.resolve(NotificationStore.self)!,
+                eventQueue: resolver.resolve(EventQueue.self)!
             )
         }
